+++ composer.json
@@ -1,33 +1,43 @@
 {
-	"name": "soapbox/laravel-formatter",
-	"type": "library",
-	"description": "A formatting library that converts data output between XML, CSV, JSON, TXT, YAML and a few others.",
-	"keywords": ["laravel", "formatter", "data", "convert", "csv", "xml", "yaml"],
+    "name": "soapbox/laravel-formatter",
+    "type": "library",
+    "description": "A formatting library that converts data output between XML, CSV, JSON, TXT, YAML and a few others.",
+    "keywords": ["laravel", "formatter", "data", "convert", "csv", "xml", "yaml"],
     "homepage": "http://github.com/SoapBox/laravel-formatter",
     "license": "MIT",
-	"version": "2.0",
-	"authors": [
-		{
-			"name": "Graham McCarthy",
-			"email": "graham@soapboxhq.com",
-			"homepage": "http://grahammccarthy.com"
-		},
-		{
-			"name": "Jaspaul Bola",
-			"email": "jaspaul.b@gamil.com",
-			"homepage": "http://jaspaulbola.com"
-		}
-	],
-	"require": {
-		"php": ">=5.4.0",
-		"league/csv": "~6.0",
-		"mustangostang/spyc": "0.5.*@dev",
-		"illuminate/support": ">=4.0"
-	},
-	"autoload": {
-		"psr-0": {
-			"SoapBox\\Formatter": "src/"
-		}
-	},
-	"minimum-stability": "dev"
+    "version": "3.0",
+    "authors": [
+        {
+            "name": "Graham McCarthy",
+            "email": "graham@soapboxhq.com",
+            "homepage": "http://grahammccarthy.com"
+        },
+        {
+            "name": "Jaspaul Bola",
+            "email": "jaspaul.b@gamil.com",
+            "homepage": "http://jaspaulbola.com"
+        }
+    ],
+    "require": {
+        "php" : ">=7.0.10",
+        "league/csv": "~9.0",
+        "mustangostang/spyc": "~0.6",
+        "illuminate/support": "5.5.x"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^5.7 || ^6.5"
+    },
+    "autoload": {
+        "psr-4": {
+            "SoapBox\\Formatter\\": "src/"
+        }
+    },
+    "extra": {
+        "laravel": {
+            "aliases": {
+                "Formatter": "SoapBox\\Formatter"
+            }
+        }
+    },
+    "minimum-stability": "stable"
 }
