--- src/SoapBox/Formatter/Parsers/JsonParser.php
@@ -1,15 +1,0 @@
-<?php namespace SoapBox\Formatter\Parsers;
-
-class JsonParser extends Parser {
-
-	private $json;
-
-	public function __construct($data) {
-		$this->json = json_decode(trim($data));
-	}
-
-	public function toArray() {
-		return (array) $this->json;
-	}
-
-}
