+++ tests/Bootstrap.php
@@ -1,5 +1,5 @@
 <?php
 
 if (!@include_once __DIR__ . '/../vendor/autoload.php') {
-	exit("Misisng projecte dependencies.");
+    exit("Misisng projecte dependencies.");
 }
