+++ tests/TestCase.php
@@ -1,17 +1,20 @@
 <?php namespace SoapBox\Formatter\Test;
 
-abstract class TestCase extends \PHPUnit_Framework_TestCase {
+abstract class TestCase extends \PHPUnit\Framework\TestCase
+{
 
-	/**
-	 * Runs prior to every test in the suite.
-	 */
-	public function setUp() {
-	}
+    /**
+     * Runs prior to every test in the suite.
+     */
+    public function setUp()
+    {
+    }
 
-	/**
-	 * Runs after every test in the suite.
-	 */
-	protected function tearDown() {
-	}
+    /**
+     * Runs after every test in the suite.
+     */
+    protected function tearDown()
+    {
+    }
 
 }
